requirements.txt
numpy==1.23.3
pandas==1.4.4
matplotlib==3.6.0
notebook==6.4.12
jupyter_nbextensions_configurator==0.5.0
hydra-core==1.2.0
torch==1.12.1
pytorch-lightning==1.7.6
clearml==1.7.0
pytest==7.1.3
pytest-cov==3.0.0

# --------- code prettifiers --------- #
flake8==5.0.4
black==22.8.0
isort==5.10.1